Deenabranch

+++ routes/web.php
@@ -31,16 +31,16 @@ Mai Section
 
 
 */
-Route::get('receptionist','UsersController@home');
-Route::get('receptionist/manage','UsersController@manageClients');
-Route::get('receptionist/approved','UsersController@approvedClients');
-Route::get('receptionist/reservations','UsersController@reservations');
-Route::get('receptionist/{id}/approve','UsersController@approve');
-Route::get('receptionist/{id}/delete','UsersController@delete');
+Route::get('receptionist','UsersController@home')->middleware('forbid-banned-user');
+Route::get('receptionist/manage','UsersController@manageClients')->middleware('forbid-banned-user');
+Route::get('receptionist/approved','UsersController@approvedClients')->middleware('forbid-banned-user');
+Route::get('receptionist/reservations','UsersController@reservations')->middleware('forbid-banned-user');
+Route::get('receptionist/{id}/approve','UsersController@approve')->middleware('forbid-banned-user');
+Route::get('receptionist/{id}/delete','UsersController@delete')->middleware('forbid-banned-user');
 //*********************************Admin*****
-Route::get('admin/clients','UsersController@showClients');
-Route::get('admin/clients/{id}/edit','UsersController@editClient');
-Route::get('admin/clients/{id}/delete','UsersController@deleteClient');
+Route::get('admin/clients','UsersController@showClients')->middleware('forbid-banned-user');
+Route::get('admin/clients/{id}/edit','UsersController@editClient')->middleware('forbid-banned-user');
+Route::get('admin/clients/{id}/delete','UsersController@deleteClient')->middleware('forbid-banned-user');
 /*
 
 
@@ -52,27 +52,27 @@ Deena Section
 Route::get(
     'managers',
     'ManagersController@index'
-)->name('managers.index');
+)->name('managers.index')->middleware('forbid-banned-user');
 Route::get(
     'receptionists',
     'ReceptionistsController@index'
-)->name('receptionists.index');
-Route::get('managers/create','ManagersController@create');
-Route::get('managers/getdata','ManagersController@getdata')->name('managers.data');
-Route::post('managers','ManagersController@store');
-Route::get('managers/{id}/edit','ManagersController@edit');
-Route::put('managers/{id}','ManagersController@update');
-Route::get('managers/{id}','ManagersController@show');
-Route::delete('managers/{id}','ManagersController@destroy');
+)->name('receptionists.index')->middleware('forbid-banned-user');
+Route::get('managers/create','ManagersController@create')->middleware('forbid-banned-user');
+Route::get('managers/getdata','ManagersController@getdata')->name('managers.data')->middleware('forbid-banned-user');
+Route::post('managers','ManagersController@store')->middleware('forbid-banned-user');
+Route::get('managers/{id}/edit','ManagersController@edit')->middleware('forbid-banned-user');
+Route::put('managers/{id}','ManagersController@update')->middleware('forbid-banned-user');
+Route::get('managers/{id}','ManagersController@show')->middleware('forbid-banned-user');
+Route::delete('managers/{id}','ManagersController@destroy')->middleware('forbid-banned-user');
 
-Route::get('receptionists/create','ReceptionistsController@create');
-Route::get('receptionists/getdata','ReceptionistsController@getdata')->name('receptionists.data');
-Route::post('receptionists','ReceptionistsController@store');
-Route::get('receptionists/{id}/edit','ReceptionistsController@edit');
-Route::put('receptionists/{id}','ReceptionistsController@update');
-Route::get('receptionists/{id}','ReceptionistsController@show');
-Route::delete('receptionists/{id}','ReceptionistsController@destroy');
-
+Route::get('receptionists/create','ReceptionistsController@create')->middleware('forbid-banned-user');
+Route::get('receptionists/getdata','ReceptionistsController@getdata')->name('receptionists.data')->middleware('forbid-banned-user');
+Route::post('receptionists','ReceptionistsController@store')->middleware('forbid-banned-user');
+Route::get('receptionists/{id}/edit','ReceptionistsController@edit')->middleware('forbid-banned-user');
+Route::put('receptionists/{id}','ReceptionistsController@update')->middleware('forbid-banned-user');
+Route::get('receptionists/{id}','ReceptionistsController@show')->middleware('forbid-banned-user');
+Route::delete('receptionists/{id}','ReceptionistsController@destroy')->middleware('forbid-banned-user');
+Route::get('receptionists/{id}/bann','ReceptionistsController@bann')->middleware('forbid-banned-user');
 /*
 
 
@@ -97,4 +97,7 @@ Route::post('rooms','RoomsController@store');
 Route::get('rooms/{id}/edit','RoomsController@edit');
 Route::patch('rooms/{id}','RoomsController@update');
 Route::delete('rooms/{id}','RoomsController@delete');
+
+Auth::routes();
+Route::get('/home', 'HomeController@index')->name('home')->middleware('forbid-banned-user');
